run_pulseまわりの責務分離
